[io] Remove call to FATAL on a socket error event with no error
Remove call to FATAL on a socket error event when we get back a null
error from the call to getsockopt with SO_ERROR, instead we just close
the socket.
(Should address the increased crash reports that apps are seeing after the FATAL call was added to see if we could get a stack trace when this condition happens).

TEST=ci
Change-Id: I39a399a7f60b218c665d30d6c43e7c18173c0e80
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/250346
Reviewed-by: Alexander Aprelev <aam@google.com>
Commit-Queue: Siva Annamalai <asiva@google.com>