[ VM / Service ] Remove required from constructor params
The required keyword means that adding a field to a class is a breaking
change, and needs a major version bump instead of a minor one. All the
fields are already nullable, and the constructors all use named
parameters, so removing this keyword is a minor (non-breaking) change.

Bug: https://github.com/dart-lang/coverage/pull/412
Bug: https://buganizer.corp.google.com/issues/236964692
Change-Id: I410d0f4359c003696570dfb11e3e2f7f179fb9ee
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/251443
Commit-Queue: Liam Appelbe <liama@google.com>
Reviewed-by: Ben Konyi <bkonyi@google.com>