Track the `VariableElement` associated with constant `DartObject`s.
The motivation for this functionality is lints which inspect the
declaration of constant values within Dart programs. By giving lints a
small amount of visibility into the intermediate values which were used
to construct constant values, it is possible for lints to validate not
just the resulting constant, but the means by which that constant was
constructed.

Change-Id: I75b3f46222ced88a0c7a7f7887de2994fcfa3f68
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/251682
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>