windows build: allow to ignore missing debug dll with env var
On Windows on Arm, when installing SDK, x64 dlls are missing. Thus, it's
impossible to build for x64 from a windows on arm machine.

In more, it can be convenient to not require those dlls to be present,
as visual studio do not install them by default (Windows SDK must be
explicitely set).

environment variable to set is "IGNORE_MISSING_DEBUG_DLL"