An initial sketch of the new refactoring support
This is very incomplete at this point, but any feedback, including
feedback about the overall design, is very welcome. I added a
placeholder for a refactoring so that I could see what the design
implies in terms of the difficulty of adding a refactoring.

I expect that specific details will change a fair bit as we start to
write some concrete refactorings, so I'm more interested in problems
with the overall design, but any comments are, of course, welcome.

Change-Id: Id80741b60542bee31fd48a0655ef8650192b58a1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/253604
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Danny Tuppeny <danny@tuppeny.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>