Fix possible raciness
wake_pause_event_handler_count_ doesn't need to be atomic,
because it is only ever accessed with pause_loop_monitor_ taken.

It is possible for WakePauseEventHandler to enter the
pause_loop_monitor_  after PauseEventHandler is done. In that case
PauseEventHandler already restored the original callback and
WakePauseEventHandler should call that instead of incrementing
wake_pause_event_handler_count_.