[analyzer] Remove the [ImportedElement] class
When running analyzer on flutter and it's transitive sources, it
consumes around 480 MB of memory.

The [ImportedElement] class is an intermediatey wrapper around an
[Element] and a boolean that is mostly false.

This CL removes the [ImportedElement] class and stores the boolean
side-information alongside the element in [PrefixScope] and
[PrefixScopeLookupResult].

This saves around 15.5 MB of memory.

TEST=ci

Change-Id: I94692a9709c70decd90c604bbdc85f5b308b3006
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/256029
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>