First draft of some process docs for the analyzer
I'm not sure that the list of "features" (that isn't the right word here)
is either correct or complete. I'm happy to fix it as part of this CL;
I'm also happy to land this as-is and let someone else fix it if that's
easier.

Change-Id: I9aa56e8eb1116478eedd0b37c91b8197e89809ca
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/256217
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>